Renamed 'kaspaperlib' to generic 'paper' library
