+++ main.go
@@ -8,7 +8,7 @@ import (
 	"github.com/pkg/errors"
 
 	"github.com/kaspanet/kaspad/domain/dagconfig"
-	"github.com/svarogg/kaspaper/kaspaperlib"
+	"github.com/karlsen-network/karlsen-paper/paper"
 )
 
 func main() {
@@ -18,7 +18,7 @@ func main() {
 	}
 	filename := os.Args[1]
 
-	wallet, err := kaspaperlib.NewAPI(&dagconfig.MainnetParams).GenerateWallet()
+	wallet, err := paper.NewAPI(&dagconfig.MainnetParams).GenerateWallet()
 	if err != nil {
 		printErrorAndExit(err)
 	}
