+++ paper/generate_password.go
@@ -1,4 +1,4 @@
-package kaspaperlib
+package paper
 
 func generatePassword() string {
 	return "password" // TODO: v v v secure random password, nobody ever will guess
